HACK: net: stmmac: Don't enable safety features if not defined by the platform
dwmac-qcom-ethqos doesn't create safety_feat_cfg, and therefore when the
hardware is detected as having asp (automotive safety package) it dies
on a NULL pointer trying to set it up.

Qualcomm didn't define this stuff downstream in the patchset, they did
this. For now leave this in place until I figure out an appropriate
definition for the structure. With the lack of documentation both in the
driver and in my hands it would be a blind guess on what's appropriate
to enable.

Signed-off-by: Andrew Halaney <ahalaney@redhat.com>