HACK: net: stmmac: dwmac-qcom-ethqos: Add EMAC3 support
Fill in the boilerplate necessary for a new platform, and in addition
add in regulator support, some different settings based on the version
being EMAC3 or not, and some settings for the particular mac2mac
configuration on the QDrive3.

The different settings should probably be cleaned up in a smarter way,
there's a lot of if statements due to them. Later commits from Qualcomm
downstream seem to do this, but I've only focused on a smaller initial
set (based on initial feedback indicating later patches were only
related to AVB, which doesn't seem to be the case unfortunately).

Likewise, the mac2mac specific conditional here should probably be a
devicetree boolean that a board sets.

Signed-off-by: Andrew Halaney <ahalaney@redhat.com>