makefile: Add -Werror support for RHEL
The -Werror flag should be enabled for RHEL.  In ARK, however, it is a bit
tricky to enable -Werror as a build bug occasionally leaks through
upstream review and into Linus' tree.  As a consequence, CONFIG_WERROR
cannot be enabled in ARK but the code to add the flag for RHEL commpiles
can be added to the top-level makefile.

This code could be placed in the top-level Makefile, around the -Werror
line added by Linus in 3fe617ccafd6 ("Enable '-Werror' by default for all
kernel builds"), but given the amount of RHEL Only code here it is best to
keep it separated in our top-level makefile.

Signed-off-by: Prarit Bhargava <prarit@redhat.com>