HACK: arm64: dts: qcom: sa8540p-ride: Add EMAC0
Add EMAC0 support. EMAC0 is hooked up to a phy that needs to be in
RGMII_SGMII mode (the media side is SGMII to a switch that doesn't support
auto-negotiation). The phy also needs to add a TX delay, so make sure to
specify the phy-mode properly.

To turn off auto-negotiation from userspace to get traffic flowing, the
following can be done:

    ethtool -s eth0 autoneg off speed 1000 duplex full

Signed-off-by: Andrew Halaney <ahalaney@redhat.com>