HACK: Reduce changes for emac0/marvell phy
I'm trying to see the minimal amount of changes to get things standing
up properly, and this works as is. Save it off.

Namely, the AUTONEG_DISABLED bit is necessary as is with the marvell phy
changes.

I'd like to see what happens if I properly describe the interrupt in the
devicetree, but baby steps, polling works ok for me.

Signed-off-by: Andrew Halaney <ahalaney@redhat.com>