net: phy: marvell: Use m88e1510_config_aneg for config_aneg
With the proper devicetree definition, this works well.

There's some things that need to be cleaned up, which I have left
comments for (namely a bad comment, and a new condition for RGMII_SGMII
mode which I still don't know how to properly identify in the
devicetree, and then also branch off of in the driver).

Signed-off-by: Andrew Halaney <ahalaney@redhat.com>