HACK: net: phy: marvell: Get things to be less invasive
Saving off this point as I think it is a good starting point to clean
things up from.

his 88EA1512 matches up with the 88E151x family (88E1510
through out here). A few things of note though.

The phy needs to be put into rgmii-txid mode to add a TX_DELAY on rgmii.

The phy needs to be put into RGMII_SGMII mode instead of the default of
copper. As far as I can tell there's no way to indicate what the "media"
side is here. Maybe a devicetree property in the phy node is appropriate
as we encode the rgmii part in the MAC on the "system" side.

The phy needs to be configured to have auto-negotiation off. I see no
way to staticly configure this in the kernel itself that isn't really
hacky. You can however configure this from userspace (one way here):
    ethtool -s eth0 autoneg off speed 1000 duplex full
According to the team that did the original board work on QNX, this is
due to the configuration of the Marvell 88Q5072 this is connected to on
the media side and is the preferred board configuration.

There's a few other comments I stumbled thru on the way that need some
cleaning up as well, which I've noted in the commits.

I'll squash this all down as appropriate now.

Signed-off-by: Andrew Halaney <ahalaney@redhat.com>