HACK: arm64: dts: qcom: sa8540p-ride: Set ethernet0 phy to RGMII_SGMII mode
This is the easiest way to do this. The alternative I can come up with
is adding a new devicetree property similar to phy-mode but called
media-mode, and latch on to it being SGMII.

This doesn't get traffic flowing though. The switch on the SGMII side
isn't setup to support auto-negotiation. This needs to be turned off in
userspace via something like:

    ethtool -s eth0 autoneg off speed 1000 duplex full

I don't see anyway to indicate this limitation in the devicetree. The
phy itself supports it, so it makes sense not to limit the phy driver
itself. Another devicetree property might be warranted to describe such
a board limitation on the media side of this chain, but since I don't
think that's acceptable (just based on the lack of anything similar that
I can find so far and the thought that this can't be an uncommon
configuration) force userspace to configure this properly.

Should be squashed into the patch adding ethernet0.

Signed-off-by: Andrew Halaney <ahalaney@redhat.com>