[redhat] New configs in drivers/iommu
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: INTEL_IOMMU_PERF_EVENTS [=y]
 Type  : bool
 Defined at drivers/iommu/intel/Kconfig:99
   Prompt: Intel IOMMU performance events
   Depends on: IOMMU_SUPPORT [=y] && INTEL_IOMMU [=y] && PERF_EVENTS [=y]
   Location:
     -> Device Drivers
       -> IOMMU Hardware Support (IOMMU_SUPPORT [=y])
         -> Support for Intel IOMMU using DMA Remapping Devices (INTEL_IOMMU [=y])
           -> Intel IOMMU performance events (INTEL_IOMMU_PERF_EVENTS [=y])

---

Cc: Jerry Snitselaar <jsnitsel@redhat.com>
Cc: Myron Stowe <mstowe@redhat.com>
Cc: Alex Williamson <alex.williamson@redhat.com>
Cc: Donald Dutile <ddutile@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>