fedora: wifi: libertas: use the LIBERTAS_THINFIRM driver
The Marvell Libertas WiFi device is quite old, primarily known to
be used on the original x86 based OLPC laptops, and likely a few
other 32bit x86 laptops. The original LIBERTAS driver uses the old
WEXT so switch to LIBERTAS_THINFIRM driver which works as well but
is more modern on the off chance there's any actual users still about.

Signed-off-by: Peter Robinson <pbrobinson@redhat.com>