HACK: stmmac irq change to handle SGMII, no info downstream on why this is done but its needed right now
unfortunately, even tho things are working, I'm maxing out at like 200 Mb/s
while downstream is getting close to the full 1 Gb/s.

It could be something in upstream, or it could be the forward port of
downstream. It is almost definitely the latter, but downstream is
way different in all respects (kernel, driver, config) than here.

Signed-off-by: Andrew Halaney <ahalaney@redhat.com>