[redhat] New configs in drivers/base
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: FW_DEVLINK_SYNC_STATE_TIMEOUT [=n]
 Type  : bool
 Defined at drivers/base/Kconfig:233
   Prompt: sync_state() behavior defaults to timeout instead of strict
   Location:
     -> Device Drivers
       -> Generic Driver Options
         -> sync_state() behavior defaults to timeout instead of strict (FW_DEVLINK_SYNC_STATE_TIMEOUT [=n])

---

 Symbol: FW_LOADER_DEBUG [=y]
 Type  : bool
 Defined at drivers/base/firmware_loader/Kconfig:29
   Prompt: Log filenames and checksums for loaded firmware
   Depends on: (CRYPTO [=y]=FW_LOADER [=y] || CRYPTO [=y]=y) && DYNAMIC_DEBUG [=y] && FW_LOADER [=y]
   Location:
     -> Device Drivers
       -> Generic Driver Options
         -> Firmware loader
           -> Firmware loading facility (FW_LOADER [=y])
             -> Log filenames and checksums for loaded firmware (FW_LOADER_DEBUG [=y])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>