[redhat] New configs in drivers/hv
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: HYPERV_VTL_MODE [=n]
 Type  : bool
 Defined at drivers/hv/Kconfig:16
   Prompt: Enable Linux to boot in VTL context
   Depends on: X86_64 [=y] && HYPERV [=m]
   Location:
     -> Device Drivers
       -> Microsoft Hyper-V guest support
         -> Microsoft Hyper-V client drivers (HYPERV [=m])
           -> Enable Linux to boot in VTL context (HYPERV_VTL_MODE [=n])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>