HACK: net: stmmac: Don't do some link stuff if pcs == SGMII
I don't understand what this change really does, but without it the link
is not coming up.

With it, and with the phy in polling mode, it does.
I'm not sure if the phy in IRQ mode would make a difference, but with
this as is now it fails to come up in IRQ mode. Should test without this
and with IRQ mode.

Good chance the serdes phy taking this thing to SGMII should be involved
here too... Need to study this some more. Downstream lumped this in the
"add sgmii support" commit with no comment on this particular part of
the change.

Signed-off-by: Andrew Halaney <ahalaney@redhat.com>