HACK: net: stmmac: dwmac-qcom-ethqos: shoehorn in 8775p support
This votes for some regulators, and smashes in a serdes.

Really the regulator stuff should be done properly (if needed, outside
the serdes).

The serdes deserves its own phy driver, which would manage the clock and
regulators it uses.

As is, for some reason this is not functioning at the expected 1G speed.
But it at least sort of works. Something from the downstream version of
this is probably missing (easy to miss, it looks a lot different down
there). If I had to guess it would be something in the serdes bit, maybe
when switching speeds, but it very well could be that MAC programming is
slightly different too.

Signed-off-by: Andrew Halaney <ahalaney@redhat.com>