HACK: Make interconnect modules builtin
Interconnect modules can be a source of pain as modules, build them in,
which also solves a issue on boot where they're not in the initramfs and
we get a dmesg storm of:

qcom a8f8800.usb: failed to get usb-ddr path: -517

aka -EPROBE_DEFER. That probably shouldn't spam as bad as it is.
It is also followed by a irq: type mismatch for irqs 12-15, which might
be an error in the error handling honestly.

Signed-off-by: Andrew Halaney <ahalaney@redhat.com>