firmware: arm_scmi: Add the SCMI performance domain
To enable support for performance scaling (DVFS) for generic devices with
the SCMI performance protocol, let's add an SCMI performance domain. This
is being modelled as a genpd provider, with support for performance scaling
through genpd's ->set_performance_state() callback.

Note that, this adds the initial support that allows consumer drivers for
attached devices, to vote for a new performance state via calling the
dev_pm_genpd_set_performance_state(). However, this should be avoided as
it's in most cases preferred to use the OPP library to vote for a new OPP
instead. The support using the OPP library is implemented from subsequent
changes.

Change-Id: I13720de301efb7897de54cc7f7564b7cf1ebe863
Signed-off-by: Ulf Hansson <ulf.hansson@linaro.org>
Patch-mainline: linux-arm-kernel @ 08/25/23, 13:26:33
Link: https://lore.kernel.org/all/20230825112633.236607-14-ulf.hansson@linaro.org/
Signed-off-by: Nikunj Kela <quic_nkela@quicinc.com>