OPP: Extend dev_pm_opp_data with OPP provider support
To allow a dynamically added OPP to be coupled with a specific OPP provider
type, let's add a new enum variable in the struct dev_pm_opp_data.
Moreover, let's add support for a DEV_PM_OPP_TYPE_GENPD type, corresponding
to genpd's performance states support.

More precisely, this allows a genpd provider to dynamically add OPPs when a
device gets attached to it, that later can be used by a consumer driver
when it needs to change the performance level for its corresponding device.

Change-Id: I612e07b67510eb5f654eb19fa66829c4a46af938
Signed-off-by: Ulf Hansson <ulf.hansson@linaro.org>
Patch-mainline: linux-arm-kernel @ 06/07/23, 14:46:25
Link: https://lore.kernel.org/all/20230607124628.157465-14-ulf.hansson@linaro.org/
Signed-off-by: Nikunj Kela <quic_nkela@quicinc.com>