firmware: arm_scmi: Add qcom hvc/shmem transport
Add a new transport channel to the SCMI firmware interface driver for
SCMI message exchange on Qualcomm virtual platforms.

The hypervisor associates an object-id also known as capability-id
with each hvc doorbell object. The capability-id is used to identify the
doorbell from the VM's capability namespace, similar to a file-descriptor.

The hypervisor, in addition to the function-id, expects the capability-id
to be passed in x1 register when HVC call is invoked.

The qcom hvc doorbell/shared memory transport uses a statically defined
shared memory region that binds with "arm,scmi-shmem" device tree node.

The function-id & capability-id are allocated by the hypervisor on bootup
and are stored in the shmem region by the firmware before starting Linux.

Currently, there is no usecase for the atomic support therefore this driver
doesn't include the changes for the same.

Change-Id: I3576fc581383a641880c7dbedec1e453839d58ee
Patch-mainline: linux-arm-kernel @ 08/11/23 10:57:19
Link: https://lore.kernel.org/all/20230811175719.28378-4-quic_nkela@quicinc.com/
Signed-off-by: Nikunj Kela <quic_nkela@quicinc.com>