firmware: arm_scmi: Align perf ops to use domain-id as in-parameter
Most scmi_perf_proto_ops are already using an "u32 domain" as an
in-parameter to indicate what performance domain we shall operate upon.
However, some of the ops are using a "struct device *dev", which means that
an additional OF parsing is needed each time the perf ops gets called, to
find the corresponding domain-id.

To avoid the above, but also to make the code more consistent, let's
replace the in-parameter "struct device *dev" with an "u32 domain". Note
that, this requires us to make some corresponding changes to the scmi
cpufreq driver, so let's do that too.

Change-Id: I54165fbc7cbf2cbd5be4aacb35ff709f64444433
Signed-off-by: Ulf Hansson <ulf.hansson@linaro.org>
Patch-mainline: linux-arm-kernel @ 08/25/23, 13:26:24
Link: https://lore.kernel.org/all/20230825112633.236607-5-ulf.hansson@linaro.org/
Signed-off-by: Nikunj Kela <quic_nkela@quicinc.com>