firmware: arm_scmi: Extend perf support with OPP from genpd providers
To enable a genpd provider to add OPPs for its attached devices,  let's
convert into using the dev_pm_opp_add_dynamic() API, in favor of the
current dev_pm_opp_add() API. This allows us to specify the frequency, the
performance level and the OPP provider type for each OPP that it may be
adding.

Moreover, to let callers of the ->device_opps_add() ops, to specify the OPP
provider let's add a new in-parameter to it.

Change-Id: I30d08c8f4661e4e189f5518f0ef58ad0565c7bd1
Signed-off-by: Ulf Hansson <ulf.hansson@linaro.org>
Patch-mainline: linux-arm-kernel @ 06/07/23, 14:46:27
Link: https://lore.kernel.org/all/20230607124628.157465-16-ulf.hansson@linaro.org/
Signed-off-by: Nikunj Kela <quic_nkela@quicinc.com>