soc: qcom: hab: Add HAB open canceling support
Give a chance for a HAB client process in user-space
to cancel a blocking HAB open operation which is in
another thread.

Once the MMID is added into the forbidden list of this
HAB client's ctx, the HAB client process will not be able
to use the habmm_socket_open service for this forbidden
MMID until the current ctx is released and a new ctx is
re-created.

So far this feature is only supposed to be used by GFX BE
case for its normal cleanup procedure for its FE exiting.

Change-Id: I215aad72de6806b78bde17b692a10a8f75874db8
Signed-off-by: Zelong Ren <quic_zeloren@quicinc.com>