OPP: Extend support for the opp-level beyond required-opps
At this point the level (performance state) for an OPP is currently limited
to be requested for a device that is attached to a PM domain.  Moreover,
the device needs to have the so called required-opps assigned to it, which
are based upon OPP tables being described in DT.

To extend the support beyond required-opps and DT, let's enable the level
to be set for all OPPs. More precisely, if the requested OPP has a valid
level let's try to request it through the device's optional PM domain, via
calling dev_pm_domain_set_performance_state().

Change-Id: Icd12381e5a7ea92a7ed5db07a287882c8a7c1409
Signed-off-by: Ulf Hansson <ulf.hansson@linaro.org>
Patch-mainline: linux-arm-kernel @ 09/25/23 15:17
Link: https://lore.kernel.org/all/20230925131715.138411-7-ulf.hansson@linaro.org/
Signed-off-by: Nikunj Kela <quic_nkela@quicinc.com>