net: stmmac: dwmac-qcom-ethqos: Hardcode the stmmac_clk rate for sa8775p
Despite being prepared and enabled, clk_get_rate() returns 0 for the
current rate of the this clock. This causes issues since calculations
around things like riwt are using this rate to translate a user's
RX coalesce settings to an appropriate value.

This shows the clk is prepared and enabled, but, its rate is reporting 0,
which messes up any driver logic that needs to use this rate to program
the IP being fed the clock. stmmaceth is stmmac_clk:

    [root@sa8775p ~]# cat /sys/kernel/debug/clk/clk_summary
                                     enable  prepare  protect
       clock                          count    count    count        rate
    ------------------------------------------------------------------------
    ...
     gcc_emac0_axi_clk                   1       1        0        0

Per Qualcomm:

    These types of clock are sourced from some NoC, in this case it
    aggre_noc and their frequencies are controlled based on the b/w
    voting on these NoCs .. So we do not have a provision to measure
    these clock's actual frequency using some API.

    I checked dumping the registers and found the actual frequencies to
    be 240MHz for both emac0/1_axi clocks.

Right now the platform can't set any coalesce values due to that with
ethtool. Since this value is necessary to program the hardware
correctly, let's accept the fact that this information must be hardcoded
and do so here.

Fixes: 8c4d92e82d50 ("net: stmmac: dwmac-qcom-ethqos: add support for emac4 on sa8775p platforms")
Suggested-by: Shazad Hussain <quic_shazhuss@quicinc.com>
Signed-off-by: Andrew Halaney <ahalaney@redhat.com>