net: stmmac: dwmac1000: convert sgmii/rgmii "pcs" to phylink
Convert dwmac1000 sgmii/rgmii "pcs" implementation to use a phylink_pcs
so we can eventually get rid of the exceptional paths that conflict
with phylink.

We do not provide a validate method to enforce auto-negotiation,
because the ethtool autonegotiation is a property of the media facing
link, not of the internal network device links.

Signed-off-by: Russell King (Oracle) <rmk+kernel@armlinux.org.uk>
Tested-by: Andrew Halaney <ahalaney@redhat.com> # sa8775p-ride
Link: https://lore.kernel.org/r/E1sautn-000tvC-7w@rmk-PC.armlinux.org.uk