SQUASH/CLEAN: net: stmmac: Use phylink instead of ps-speed to program speed and RAL
The SGMIRAL bit is used to force the SGMII RAL (rate adaptation layer)
to operate in the speed configured in the MAC configuration register's
speed and port select bits instead of what the PCS determines. This is
intended for MAC2MAC (aka fixed-link) setups that can't get this info
from the PCS.

Instead of using a devicetree property (snps,ps-speed) use the phylink
provided information to figure out when to set the bit.

TODO: verify pre-patch flow below!
Currently, snps,ps-speed is used to program this bit. As part of
stmmac_open() the MAC_CTRL register is programmed with the speed
indicated in that property. The SGMIRAL bit is then set in
(TODO when is this set?) to indicate that the MAC_CTRL setting should
be honored instead of what the PCS determines.

This is unnecessary, let's use phylink's mac_link_up() and pcs_config()
callbacks to figure out what to program here. mac_link_up() already
takes care of programming the speed (unconditionally) etc into MAC_CTRL.
TODO: mention Port Select probably based on Serge's description
pcs_config() now looks at neg_mode to determine if the settings are
provided inband or not, and sets the SGMIRAL bit if they are not.

This removes the need for snps,ps-speed altogether. For that reason
deprecate the property and remove associated variables.

In addition, dwmac-qcom-ethqos plays with the PCS independent of
phylink. Stop doing that. Everything should be done in the phylink PCS
callbacks now.

Signed-off-by: Andrew Halaney <ahalaney@redhat.com>